Add more sort fields as necessary to make sure the sorted feeds make sense to a human.
